src/backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dj_database_url==2.1.0
matplotlib==3.8.2
psycopg2==2.9.3
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.2
pandas==2.1.3
scikit-survival==0.22.1
shap==0.44.0
xgboost==1.7.6